Docker/requirements.txt
flask
flask-smorest
python-dotenv
numpy